+++ position.h
@@ -5,11 +5,11 @@
 class position
 {
 public:
-    position(double x, double y);
+    position(int x, int y);
     position(); // par default (0,0)
-    double valeurX() const;
-    double valeurY() const;
-    void changePosition(double x, double y);
+    int valeurX() const;
+    int valeurY() const;
+    void changePosition(int x, int y);
     void affiche(std::ostream& ost) const;
     void lit(std::istream& ist);
     bool auDessus(const position& pos) const;
@@ -20,8 +20,8 @@ public:
 
 
 private:
-    double d_x;
-    double d_y;
+    int d_x;
+    int d_y;
 };
 std::ostream& operator<<(std::ostream& ost, const position& pos);
 std::istream& operator>>(std::istream& ist, position& pos);
